src/backend/requirements.txt
fastapi
uvicorn
requests
google-cloud-monitoring
google-cloud-logging
